Refactor : CustomException 사용해서 예외 처리

+++ src/main/java/com/example/miniproject/service/UserService.java
@@ -3,7 +3,10 @@ package com.example.miniproject.service;
 import com.example.miniproject.config.jwt.JwtUtil;
 import com.example.miniproject.dto.LoginRequestDto;
 import com.example.miniproject.dto.SignupRequestDto;
+import com.example.miniproject.dto.http.ResponseMessage;
+import com.example.miniproject.dto.http.StatusCode;
 import com.example.miniproject.entity.User;
+import com.example.miniproject.exception.CustomException;
 import com.example.miniproject.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -30,14 +33,14 @@ public class UserService {
         String password = loginRequestDto.getPassword();
 
         User user = userRepository.findByUserId(userId).orElseThrow(
-                () -> new IllegalArgumentException("가입하지 않은 회원입니다."));// 예외처리 해주기
+                () -> new CustomException(ResponseMessage.NOT_FOUND_USER, StatusCode.UNAUTHORIZED));// 예외처리 해주기
 
         System.out.println(password);
         System.out.println(user.getPassword());
 
         System.out.println(passwordEncoder.encode(password));
         if (!passwordEncoder.matches(password, user.getPassword())) {
-            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+            throw new CustomException(ResponseMessage.NOT_FOUND_USER, StatusCode.UNAUTHORIZED);
         }
 
         String accessToken = jwtUtil.createAccessToken(user.getUserId());
@@ -53,27 +56,22 @@ public class UserService {
     @Transactional(readOnly = true)
     public void checkId(String userId) {
         if (userRepository.existsByUserId(userId)) {
-            throw new IllegalStateException("이미 사용중인 아이디입니다.");
+            throw new CustomException(ResponseMessage.ALREADY_ENROLLED_USER, StatusCode.CONFLICT);
         }
     }
 
     @Transactional
     public void logout(HttpServletRequest request) {
-        if (request.getHeader(JwtUtil.REFRESHTOKEN_HEADER) == null) {
-            throw new IllegalStateException("잘못된 접근입니다");
-        }
-
-        if (redisService.getValues(request.getHeader(JwtUtil.REFRESHTOKEN_HEADER)) != null) {
+        if (request.getHeader(JwtUtil.REFRESHTOKEN_HEADER) != null && redisService.getValues(request.getHeader(JwtUtil.REFRESHTOKEN_HEADER)) != null) {
             redisService.deleteValues(request.getHeader(JwtUtil.REFRESHTOKEN_HEADER));
-        } else {
-            throw new IllegalStateException("잘못된 접근입니다");
         }
-
+        throw new CustomException(ResponseMessage.WRONG_ACCESS, StatusCode.BAD_REQUEST);
     }
 
     //회원가입
     @Transactional
     public void signup(SignupRequestDto signupRequestDto) {
+        // 지훈님 수정 후 exception customizing 필요
         String userId = signupRequestDto.getUserId();
         String password = passwordEncoder.encode(signupRequestDto.getPassword());
         String nickname = signupRequestDto.getNickname();
